Refactor: Refactor si domain and gu domain
