+++ server/src/main/java/project/manager/server/domain/region/Si.java
@@ -1,21 +1,19 @@
-package project.manager.server.domain;
+package project.manager.server.domain.region;
 
 import jakarta.persistence.*;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-import org.hibernate.annotations.DynamicUpdate;
 
 @Entity
 @Getter
 @NoArgsConstructor
 @Table(name = "SI_TB")
-@DynamicUpdate
 public class Si {
     @Id
-    @Column(name = "id", nullable = false,updatable = false)
+    @Column(name = "id", nullable = false)
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
 
-    @Column(name = "name",nullable = false)
-    private String name;
+    @Column(name = "si", nullable = false, length = 100)
+    private String si;
 }
